+++ lib/src/time/calendar/calendar.dart
@@ -1,10 +1,16 @@
 library time.calendar;
 
 import 'package:date/date.dart';
+import 'package:elec/src/time/calendar/calendars/federal_holidays_calendar.dart';
+import 'package:elec/src/time/calendar/calendars/nerc_calendar.dart';
 import 'holiday.dart';
 
 /// A holiday calendar
 abstract class Calendar {
+  static final federalHolidays = FederalHolidaysCalendar();
+  static final nerc = NercCalendar();
+
+  bool isHoliday3(int year, int month, int day);
   bool isHoliday(Date date);
   HolidayType getHolidayType(Date date);