+++ lib/src/time/calendar/calendars/ma_state_holiday_calendar.dart
@@ -38,6 +38,12 @@ class MaStateHolidayCalendar extends Calendar {
     return res;
   }
 
+  @override
+  bool isHoliday3(int year, int month, int day) {
+    // TODO: implement isHoliday3
+    throw UnimplementedError();
+  }
+
 }