+++ lib/src/time/calendar/calendars/nerc_calendar.dart
@@ -13,18 +13,7 @@ import '../calendar.dart';
 
 /// NERC Calendar
 class NercCalendar extends Calendar {
-  static final Holiday _newYear = NewYear();
-  static final Holiday _memorialDay = MemorialDay();
-  static final Holiday _independenceDay = IndependenceDay();
-  static final Holiday _laborDay = LaborDay();
-  static final Holiday _thanksgiving = Thanksgiving();
-  static final Holiday _christmas = Christmas();
-
   late HolidayType _holidayType;
-
-  /// Store all the holidays for one year in a Map
-  /// year -> {(month,day)} as we don't care of the type of the holiday.
-  final _holidayCache = LruMap<int, Set<Tuple2<int, int>>>(maximumSize: 50);
 
   @override
   HolidayType getHolidayType(Date date) {
@@ -55,35 +44,22 @@ class NercCalendar extends Calendar {
     return _holidayType;
   }
 
+  @Deprecated('Use isHoliday3.')
   @override
-  bool isHoliday(Date date) {
-    final year = date.year;
-    if (!_holidayCache.containsKey(year)) {
-      _addYearToCache(year);
-    }
+  bool isHoliday(Date date) => isHoliday3(date.year, date.month, date.day);
 
-    if (_holidayCache[year]!.contains(Tuple2(date.month, date.day))) {
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  /// Add one year to the cache
-  void _addYearToCache(int year) {
-    var newYear = _newYear.forYear(year, location: UTC)!;
-    var memorialDay = _memorialDay.forYear(year, location: UTC)!;
-    var indDay = _independenceDay.forYear(year, location: UTC)!;
-    var laborDay = _laborDay.forYear(year, location: UTC)!;
-    var thanksDay = _thanksgiving.forYear(year, location: UTC)!;
-    var christDay = _christmas.forYear(year, location: UTC)!;
-    _holidayCache[year] = {
-      Tuple2(newYear.month, newYear.day),
-      Tuple2(memorialDay.month, memorialDay.day),
-      Tuple2(indDay.month, indDay.day),
-      Tuple2(laborDay.month, laborDay.day),
-      Tuple2(thanksDay.month, thanksDay.day),
-      Tuple2(christDay.month, christDay.day),
+  /// Note: the implementation of [isHoliday3] is about 30% faster than
+  /// [isHoliday].
+  @override
+  bool isHoliday3(int year, int month, int day) {
+    return switch (month) {
+      1 => Holiday.newYear.isDate3(year, month, day) ? true : false,
+      5 => Holiday.memorialDay.isDate3(year, month, day) ? true : false,
+      7 => Holiday.independenceDay.isDate3(year, month, day) ? true : false,
+      9 => Holiday.laborDay.isDate3(year, month, day) ? true : false,
+      11 => Holiday.thanksgiving.isDate3(year, month, day) ? true : false,
+      12 => Holiday.christmas.isDate3(year, month, day) ? true : false,
+      _ => false,
     };
   }
 }