+++ lib/src/time/calendar/calendars/nh_state_holiday_calendar.dart
@@ -38,6 +38,14 @@ class NhStateHolidayCalendar extends Calendar {
     return res;
   }
 
+  @override
+  bool isHoliday3(int year, int month, int day) {
+    // TODO: implement isHoliday3
+    throw UnimplementedError();
+  }
+
+
+
 }