+++ lib/src/time/calendar/calendars/ri_state_holiday_calendar.dart
@@ -47,6 +47,12 @@ class RiStateHolidayCalendar extends Calendar {
     return res;
   }
 
+  @override
+  bool isHoliday3(int year, int month, int day) {
+    // TODO: implement isHoliday3
+    throw UnimplementedError();
+  }
+
 }