+++ lib/src/time/calendar/holiday_utils.dart
@@ -16,9 +16,13 @@ Map<Date, String> assignHolidays(List<Date> dates, Calendar calendar) {
 /// Make a holiday if you know the month, week of the month, and weekday
 Date makeHoliday(
     int year, int month, int weekOfMonth, int weekday, {required Location location}) {
+  var day = dayOfMonthHoliday(year, month, weekOfMonth, weekday);
+  return Date(year, month, day, location: location);
+}
+
+int dayOfMonthHoliday(int year, int month, int weekOfMonth, int weekday) {
   var wdayBom = DateTime(year, month, 1).weekday;
   var inc = weekday - wdayBom;
   if (inc < 0) inc += 7;
-  return Date(year, month, 7 * (weekOfMonth - 1) + inc + 1,
-    location: location);
+  return 7 * (weekOfMonth - 1) + inc + 1;
 }