+++ lib/src/time/calendar/holidays/thanksgiving.dart
@@ -15,7 +15,17 @@ class Thanksgiving extends Holiday {
     return makeHoliday(year, 11, 4, DateTime.thursday, location: location);
   }
 
+  @Deprecated('Replace with isDate3')
   @override
   bool isDate(Date date) =>
       forYear(date.year, location: date.location) == date;
+
+
+  bool isDate3(int year, int month, int day) {
+    if (month != 11) return false;
+    if (day < 22 || day > 28) return false;
+    var dayOfMonth = dayOfMonthHoliday(year, 11, 4, DateTime.thursday);
+    if (dayOfMonth == day) return true;
+    return false;
+  }
 }