+++ test/time/calendar/calendar_test.dart
@@ -1,6 +1,7 @@
 library calendar_test;
 
 import 'package:elec/elec.dart';
+import 'package:elec/time.dart';
 import 'package:test/test.dart';
 import 'package:date/date.dart';
 import 'package:elec/src/time/calendar/holiday.dart';
@@ -14,7 +15,7 @@ import 'package:timezone/timezone.dart';
 void tests() {
   group('Test calendar', () {
     test('first business day of the month', () {
-      var calendar = NercCalendar();
+      var calendar = Calendar.nerc;
       expect(
           calendar.firstBusinessDate(Month.utc(2019, 9)), Date.utc(2019, 9, 3));
       expect(
@@ -23,8 +24,54 @@ void tests() {
           calendar.firstBusinessDate(Month.utc(2020, 2)), Date.utc(2020, 2, 3));
     });
 
+    test('NERC Calendar', () {
+
+      var days = [
+        //
+        (year: 2018, month: 1, day: 1),
+        (year: 2018, month: 5, day: 28),
+        (year: 2018, month: 7, day: 4),
+        (year: 2018, month: 9, day: 3),
+        (year: 2018, month: 11, day: 22),
+        (year: 2018, month: 12, day: 25),
+        //
+        (year: 2021, month: 1, day: 1),
+        (year: 2021, month: 5, day: 31),
+        (year: 2021, month: 7, day: 5),
+        (year: 2021, month: 9, day: 6),
+        (year: 2021, month: 11, day: 25),
+        (year: 2021, month: 12, day: 25),
+        //
+        (year: 2022, month: 1, day: 1),
+        (year: 2022, month: 5, day: 30),
+        (year: 2022, month: 7, day: 4),
+        (year: 2022, month: 9, day: 5),
+        (year: 2022, month: 11, day: 24),
+        (year: 2022, month: 12, day: 26),
+        //
+        (year: 2023, month: 1, day: 2),
+        (year: 2023, month: 5, day: 29),
+        (year: 2023, month: 7, day: 4),
+        (year: 2023, month: 9, day: 4),
+        (year: 2023, month: 11, day: 23),
+        (year: 2023, month: 12, day: 25),
+        //
+        (year: 2027, month: 1, day: 1),
+        (year: 2027, month: 5, day: 31),
+        (year: 2027, month: 7, day: 5),
+        (year: 2027, month: 9, day: 6),
+        (year: 2027, month: 11, day: 25),
+        (year: 2027, month: 12, day: 25),
+      ];
+      expect(Calendar.nerc.isHoliday3(2023, 1, 1), false);
+      for (var date in days) {
+        expect(Calendar.nerc.isHoliday3(date.year, date.month, date.day), true);
+        // expect(Calendar.nerc.isHoliday(Date.utc(date.year, date.month, date.day)), true);
+      }
+    });
+
     test('NERC Holidays 2018', () {
-      var calendar = NercCalendar();
+      var calendar = Calendar.nerc;
       expect(calendar.isHoliday(Date.utc(2018, 1, 1)), true);
       expect(
           calendar.getHolidayType(Date.utc(2018, 1, 1)), HolidayType.newYear);
@@ -104,11 +151,8 @@ void tests() {
 }
 
 /// See how fast isHoliday(date) is.
-/// For 10 years repeated 24 times it takes:
-///  - 80 ms without caching
-///  - 32 ms with LRU Map caching
-///  - 32 ms with a regular Map
-///  -  4 ms just iterating
+/// For 10 years repeated 25 times it takes:
+///  - 33 ms without caching.
 /// FWIW: A Rust implementation takes 0.7 ms (no caching).
 void speedTest() {
   var location = getLocation('America/New_York');
@@ -119,7 +163,9 @@ void speedTest() {
   var sw = Stopwatch()..start();
   for (var i = 0; i < 25; i++) {
     for (var date in days) {
+      // if (calendar.isHoliday3(date.year, date.month, date.day)) {
       if (calendar.isHoliday(date)) {
+      //   print(date);
         count++;
       }
     }
@@ -127,6 +173,7 @@ void speedTest() {
   sw.stop();
   print('Time spent: ${sw.elapsedMilliseconds} millis');
   print('Count of holidays: $count');
+  assert(count == 1500);
 }
 
 void main() {