+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## TODO
 
+## 2.1.0, release 2023-07-07
+- Made NercCalendar isHoliday about 30% faster.  Removed the use of a cache (no speed 
+  improvement).
+- Using calendar.isHoliday3 in bucket definitions made bucket speedTest 50% faster, 
+  went from 125ms to 65ms.
+
 ## Release 2023-05-31
 - Bump lower version of sdk to 3.0.2
 - Bump up packages
-
 
 ## Release 2023-05-29
 - Bump sdk upper limit to 4.0.0